Pass middleware separately, so modifySite can interject later.
Essentially, if you are adding middleware specifically for testing
particular requests, it's possible that you want to be operating on
requests that have already passed through the rest of the
middleware (for ex, if you have a middleware handling session
information).

With this change, we keep a stack of middleware to apply, and
`modifySite` adds another element to it. When we run handlers, we apply
all the middleware.

Note: This doesn't affect things for `eval`, which may be a problem - it
will be operating on a ctxt with a request that has not passed through
middleware. On the other hand, eval isn't even in the context of a
request, so it's a little bit strange to depend on request transforming
middleware.